Move NewMoney helper func to utility.go
