+++ utility.go
@@ -8,10 +8,12 @@ import (
 	"github.com/form3tech-oss/jwt-go"
 	gzerrors "github.com/gazebo-web/gz-go/v8/errors"
 	"google.golang.org/genproto/googleapis/type/datetime"
+	"google.golang.org/genproto/googleapis/type/money"
 	"google.golang.org/protobuf/types/known/durationpb"
 	htmlTemplate "html/template"
 	"io"
 	"log"
+	"math"
 	"math/rand"
 	"net/http"
 	"net/url"
@@ -448,5 +450,17 @@ func NewDateTime(t time.Time) *datetime.DateTime {
 		Seconds:    int32(t.Second()),
 		Nanos:      int32(t.Nanosecond()),
 		TimeOffset: &datetime.DateTime_UtcOffset{UtcOffset: durationpb.New(time.Duration(offset) * time.Second)},
+	}
+}
+
+// NewMoney converts the given cents into a money.Money value.
+// NOTE: money.Money is a protobuf message.
+func NewMoney(currency string, cents int64) *money.Money {
+	u := cents / 100
+	n := int32(cents-(u*100)) * int32(math.Pow10(7))
+	return &money.Money{
+		CurrencyCode: currency,
+		Units:        u,
+		Nanos:        n,
 	}
 }
